+++ src/main/java/it/eng/idsa/dataapp/configuration/SecureRestTemplateCustomizer.java
@@ -14,22 +14,19 @@ import org.apache.http.ssl.SSLContexts;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.web.client.RestTemplateCustomizer;
 import org.springframework.http.client.ClientHttpRequestFactory;
 import org.springframework.http.client.HttpComponentsClientHttpRequestFactory;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestTemplate;
 
-
 /**
  * Used for Customized TLS security, to avoid PKIX error.
  * When using RestTemplate call restTemplateBuilder.build().
  * 
- * @see it.eng.idsa.dataapp.service.impl.ProxyServiceImpl#ProxyServiceImpl(org.springframework.boot.web.client.RestTemplateBuilder, ECCProperties, it.eng.idsa.dataapp.service.RecreateFileService, String, String, Boolean, Boolean)
+ * @see it.eng.idsa.dataapp.service.impl.ProxyServiceImpl#ProxyServiceImpl(org.springframework.boot.web.client.RestTemplateBuilder, it.eng.idsa.dataapp.configuration.ECCProperties, it.eng.idsa.dataapp.service.RecreateFileService, java.util.Optional, java.lang.String, java.lang.String, java.lang.Boolean, java.lang.Boolean)
  */
 @Component
-@ConditionalOnProperty(name = "application.ecc.protocol", havingValue = "https")
 public class SecureRestTemplateCustomizer implements RestTemplateCustomizer {
 
 	private static final Logger logger = LoggerFactory.getLogger(SecureRestTemplateCustomizer.class);
@@ -45,7 +42,7 @@ public class SecureRestTemplateCustomizer implements RestTemplateCustomizer {
 		HttpClient httpClient;
 		try {
 			sslcontextBuilder.loadTrustMaterial(null, (cert, auth) -> true);
-			
+
 			SSLConnectionSocketFactory sslConnectionSocketFactory = new SSLConnectionSocketFactory(
 					sslcontextBuilder.build(), (HostnameVerifier) NoopHostnameVerifier.INSTANCE);
 			httpClient = HttpClients.custom()