+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.3.1] - 2023-07-25
+
+### Changed
+
+ - Hardcoded to use TLS communication only
+ - Property application.security.username deleted, instead the fixed value idsUser is used
+ - Dockerfile improvements
+ - SSL certificate update
+ 
 ## [0.3.0] - 2023-07-17
 
 ### Changed
@@ -8,7 +17,7 @@ All notable changes to this project will be documented in this file.
  - Add Certification check
  - Add checkSum verification
  - Property application.verifyCheckSum required for checkSum verification
- 
+
   
 ## [0.2.8] - 2023-06-20